feat: access to NetworkController in KafkaExecutionContext
A NetworkController interface is added to KafkaExecutionContext. It can
be used from inside a Policy to stop/resume the acceptance of new
requests.
It has been introduced for Kafka Quota Policy to handle throttling: when
we detect the quota limit has been reached, the GW will stop accepting
requests until the throttling time has passed.